refactor(cardano-services): align Asset provider - restructure Asset directory to be aligned with all other providers
