+++ packages/cardano-services/src/Asset/DbSyncAssetProvider/types.ts
@@ -1,0 +1,18 @@
+export interface LastMintTxModel {
+  tx_hash: Buffer;
+}
+
+export interface MultiAssetModel {
+  fingerprint: string;
+  id: string;
+}
+
+export interface MultiAssetHistoryModel {
+  hash: Buffer;
+  quantity: string;
+}
+
+export interface MultiAssetQuantitiesModel {
+  count: string;
+  sum: string;
+}
