+++ packages/cardano-services/src/Asset/index.ts
@@ -1,7 +1,6 @@
-export * from './AssetBuilder';
 export * from './AssetHttpService';
 export * from './CardanoTokenRegistry';
 export * from './StubTokenMetadataService';
-export * from './DbSyncAssetProvider';
 export * from './DbSyncNftMetadataService';
+export * from './DbSyncAssetProvider';
 export * from './types';
