+++ packages/cardano-services/src/Asset/types.ts
@@ -29,22 +29,3 @@ export interface TokenMetadataService extends Shutdown {
    */
   getTokenMetadata(assetIds: Cardano.AssetId[]): Promise<(Asset.TokenMetadata | null)[]>;
 }
-
-export interface LastMintTxModel {
-  tx_hash: Buffer;
-}
-
-export interface MultiAssetModel {
-  fingerprint: string;
-  id: string;
-}
-
-export interface MultiAssetHistoryModel {
-  hash: Buffer;
-  quantity: string;
-}
-
-export interface MultiAssetQuantitiesModel {
-  count: string;
-  sum: string;
-}
